+++ internal/provider/data_source_iosxe_crypto_ikev2_profile_test.go
@@ -47,6 +47,11 @@ func TestAccDataSourceIosxeCryptoIKEv2Profile(t *testing.T) {
 	checks = append(checks, resource.TestCheckResourceAttr("data.iosxe_crypto_ikev2_profile.test", "dpd_retry", "2"))
 	checks = append(checks, resource.TestCheckResourceAttr("data.iosxe_crypto_ikev2_profile.test", "dpd_query", "periodic"))
 	checks = append(checks, resource.TestCheckResourceAttr("data.iosxe_crypto_ikev2_profile.test", "lifetime", "28800"))
+	checks = append(checks, resource.TestCheckResourceAttr("data.iosxe_crypto_ikev2_profile.test", "authentication_local_rsa_sig", "true"))
+	checks = append(checks, resource.TestCheckResourceAttr("data.iosxe_crypto_ikev2_profile.test", "authentication_remote_rsa_sig", "true"))
+	checks = append(checks, resource.TestCheckResourceAttr("data.iosxe_crypto_ikev2_profile.test", "match_certificate_maps.0", "map1"))
+	checks = append(checks, resource.TestCheckResourceAttr("data.iosxe_crypto_ikev2_profile.test", "pki_trustpoints.0.name", "myCA"))
+	checks = append(checks, resource.TestCheckResourceAttr("data.iosxe_crypto_ikev2_profile.test", "pki_trustpoints.0.uses", "sign"))
 	checks = append(checks, resource.TestCheckResourceAttr("data.iosxe_crypto_ikev2_profile.test", "config_exchange_request", "false"))
 	resource.Test(t, resource.TestCase{
 		PreCheck:                 func() { testAccPreCheck(t) },
@@ -80,6 +85,14 @@ resource "iosxe_yang" "PreReq1" {
 	}
 }
 
+resource "iosxe_yang" "PreReq2" {
+	path = "/Cisco-IOS-XE-native:native/crypto/Cisco-IOS-XE-crypto:pki/trustpoint[id=myCA]"
+	attributes = {
+		"id" = "myCA"
+		"enrollment/enrollment-method/selfsigned" = ""
+	}
+}
+
 `
 
 // End of section. //template:end testPrerequisites
@@ -107,8 +120,15 @@ func testAccDataSourceIosxeCryptoIKEv2ProfileConfig() string {
 	config += `	dpd_retry = 2` + "\n"
 	config += `	dpd_query = "periodic"` + "\n"
 	config += `	lifetime = 28800` + "\n"
+	config += `	authentication_local_rsa_sig = true` + "\n"
+	config += `	authentication_remote_rsa_sig = true` + "\n"
+	config += `	match_certificate_maps = ["map1"]` + "\n"
+	config += `	pki_trustpoints = [{` + "\n"
+	config += `		name = "myCA"` + "\n"
+	config += `		uses = "sign"` + "\n"
+	config += `	}]` + "\n"
 	config += `	config_exchange_request = false` + "\n"
-	config += `	depends_on = [iosxe_yang.PreReq0, iosxe_yang.PreReq1, ]` + "\n"
+	config += `	depends_on = [iosxe_yang.PreReq0, iosxe_yang.PreReq1, iosxe_yang.PreReq2, ]` + "\n"
 	config += `}` + "\n"
 
 	config += `